requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.9.0.80
numpy<2.0.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.4
PyPDF2==3.0.1
openai==1.3.0
PyMuPDF==1.23.8
python-docx==0.8.11
